1.设计 RSF 配置文件，以及默认值。 2.ProtocolStatus 值，取消必须使用枚举类型，改为 short 数值类型。 3.ServiceMetaData 删除 supportAsync 同步异步配置。同步异步是客户端决定的。 4.Accepted 包的发送，改为 work 线程当成功放入 task 任务队列之后发送。 如果放入失败则发送 ChooseOther 包不会。

+++ hasor-rsf/src/main/java/net/hasor/rsf/_test/ClientHandler.java
@@ -69,15 +69,15 @@ public class ClientHandler extends ChannelInboundHandlerAdapter {
     public void channelRead(ChannelHandlerContext ctx, Object msg) throws Exception {
         ResponseMsg response = (ResponseMsg) msg;
         //
-        if (response.getStatus() == ProtocolStatus.Accepted)
+        if (response.getStatus() == ProtocolStatus.Accepted.shortValue())
             acceptedCount++;
-        else if (response.getStatus() == ProtocolStatus.ChooseOther)
+        else if (response.getStatus() == ProtocolStatus.ChooseOther.shortValue())
             chooseOtherCount++;
-        else if (response.getStatus() == ProtocolStatus.OK)
+        else if (response.getStatus() == ProtocolStatus.OK.shortValue())
             okCount++;
-        else if (response.getStatus() == ProtocolStatus.SerializeError)
+        else if (response.getStatus() == ProtocolStatus.SerializeError.shortValue())
             serializeError++;
-        else if (response.getStatus() == ProtocolStatus.RequestTimeout)
+        else if (response.getStatus() == ProtocolStatus.RequestTimeout.shortValue())
             requestTimeout++;
         else {
             int a = 0;
